#801: Fix many bugs and stabilize the new jdd page
Better messages. More disabled icons. Fix on buttons state depending on the state of the submission and export file. Reload of the page on DEE generation or cancellation. Fix on short title not correctly displayed. Fix when jdd is created and no submission is made directly after : page is now on stable state. Nb data was displayed even when a DSR was deleted : this is corrected.